Add first reducer
