+++ package.json
@@ -28,6 +28,7 @@
     "react-codemirror2": "^5.1.0",
     "react-dom": "^16.4.1",
     "react-measure": "^2.0.2",
+    "react-redux": "^5.0.7",
     "react-scripts": "1.1.4",
     "redux": "^4.0.0"
   },
