+++ src/exports.js
@@ -1,2 +1,3 @@
 export { VisualizationRunner } from './visualizationRunner/index';
 export { CodeEditor } from './visualizationEditor/index';
+export { reducers } from './reducers/index';
