+++ src/index.js
@@ -1,8 +1,1 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { TestingApp } from './testingApp';
-
-import 'codemirror/lib/codemirror.css';
-import './css/ubuntu.css';
-
-ReactDOM.render(<TestingApp />, document.getElementById('root'));
+import './testingApp/index';
