+++ src/reducers/editor.js
@@ -1,0 +1,10 @@
+export const editor = (state = {}, action) => {
+  switch (action.type) {
+    case 'INITIALIZE_FILES':
+      return Object.assign({}, state, {
+        files: action.files
+      });
+    default:
+      return state;
+  }
+}
