+++ src/reducers/index.js
@@ -1,0 +1,3 @@
+import { editor } from './editor';
+
+export const reducers = { editor };
