+++ src/testingApp/index.js
@@ -1,19 +1,19 @@
-import React, { Component } from 'react';
-import { VisualizationRunner, CodeEditor } from '../exports';
-import { files } from './files';
+import React from 'react';
+import { render } from 'react-dom';
+import { TestingApp } from './testingApp';
 
-export class TestingApp extends Component {
-  render() {
-    return (
-      <div>
-        <VisualizationRunner
-          width={960}
-          height={500}
-          files={files}
-        />
-        <div>Hello</div>
-        <CodeEditor />
-      </div>
-    );
-  }
-}
+import 'codemirror/lib/codemirror.css';
+import '../css/ubuntu.css';
+
+import { Provider } from 'react-redux'
+import { createStore } from 'redux'
+import rootReducer from './rootReducer'
+
+const store = createStore(rootReducer)
+
+render(
+  <Provider store={store}>
+    <TestingApp />
+  </Provider>,
+  document.getElementById('root')
+);
