+++ src/testingApp/rootReducer.js
@@ -1,0 +1,4 @@
+import { combineReducers } from 'redux';
+import { reducers } from '../exports';
+const { editor } = reducers;
+export default combineReducers({ editor });
